% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantiles.R
\name{hd.ineq}
\alias{hd.ineq}
\title{Upper-to-lower quantile ratio as a simple measure of inequality}
\usage{
hd.ineq(x, p = 0.1)
}
\arguments{
\item{x}{Numeric vector with positive lower quantile.}

\item{p}{Lower quantile probability, between 0 and 0.5. Default = 0.1.}
}
\value{
Numeric upper-to-lower quantile ratio.
}
\description{
Computes Q(1 - p) / Q(p), such as the 90/10 ratio.
}
