% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantiles.R
\name{hd.iqr}
\alias{hd.iqr}
\title{Interquantile range}
\usage{
hd.iqr(x, p = 0.25)
}
\arguments{
\item{x}{Numeric vector.}

\item{p}{Lower quantile probability, between 0 and 0.5. Default = 0.25.}
}
\value{
Numeric interquantile range.
}
\description{
Computes Q(1 - p) - Q(p).
}
