% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantiles.R
\name{hd.kurt}
\alias{hd.kurt}
\title{Moors quantile kurtosis}
\usage{
hd.kurt(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
Numeric Moors quantile kurtosis.
}
\description{
Computes the octile-based kurtosis measure of Moors (1988).
}
\references{
Moors, J. J. A. (1988). A quantile alternative for kurtosis.
}
