% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantiles.R
\name{hd.lts}
\alias{hd.lts}
\title{Lower tail spread}
\usage{
hd.lts(x, q.C = 0.5, q.L = 0.1)
}
\arguments{
\item{x}{Numeric vector.}

\item{q.C}{Central quantile. Default = 0.5.}

\item{q.L}{Lower quantile. Default = 0.1.}
}
\value{
Numeric lower tail spread.
}
\description{
Computes the difference between a central and lower quantile.
}
