% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantiles.R
\name{hd.ltw}
\alias{hd.ltw}
\title{Left tail weight}
\usage{
hd.ltw(x, p = 0.25)
}
\arguments{
\item{x}{Numeric vector.}

\item{p}{Quantile probability between 0 and 0.5. Default = 0.25.}
}
\value{
Numeric left tail weight.
}
\description{
Computes a quantile-based measure of left tail weight.
}
