% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantiles.R
\name{hd.niqr}
\alias{hd.niqr}
\title{Normalised interquantile range}
\usage{
hd.niqr(x, p = 0.25, q.L = 0.1, q.U = 0.9)
}
\arguments{
\item{x}{Numeric vector.}

\item{p}{Lower quantile probability, between 0 and 0.5. Default = 0.25.}

\item{q.L}{Lower reference quantile. Default = 0.1.}

\item{q.U}{Upper reference quantile. Default = 0.9.}
}
\value{
Numeric normalised interquantile range.
}
\description{
Computes (Q(1 - p) - Q(p)) / (Q(q.U) - Q(q.L)).
}
