% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantiles.R
\name{hd.qcd}
\alias{hd.qcd}
\title{Quantile coefficient of dispersion}
\usage{
hd.qcd(x, p = 0.25)
}
\arguments{
\item{x}{Numeric vector with positive quantiles.}

\item{p}{Lower quantile probability, between 0 and 0.5. Default = 0.25.}
}
\value{
Numeric quantile coefficient of dispersion.
}
\description{
Computes (Q(1 - p) - Q(p)) / (Q(1 - p) + Q(p)).
}
