% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantiles.R
\name{hd.rcv}
\alias{hd.rcv}
\title{Robust coefficient of variation}
\usage{
hd.rcv(x)
}
\arguments{
\item{x}{Numeric vector with a nonzero median.}
}
\value{
Numeric robust coefficient of variation.
}
\description{
Computes 0.75 * (Q(0.75) - Q(0.25)) / Q(0.5).
}
