% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcoxmod.R
\name{hd}
\alias{hd}
\title{Harrell-Davis estimate of a quantile}
\usage{
hd(x, q = 0.5)
}
\arguments{
\item{x}{Numeric data vector.}

\item{q}{Desired quantile, between 0 and 1.}
}
\value{
A numeric Harrell-Davis quantile estimate.
}
\description{
Harrell-Davis estimate of a quantile
}
\references{
Wilcox, Rand R. 2022. Introduction to Robust Estimation and Hypothesis Testing. 5th edn.
}
