% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantiles.R
\name{hd.rtw}
\alias{hd.rtw}
\title{Right tail weight}
\usage{
hd.rtw(x, p = 0.75)
}
\arguments{
\item{x}{Numeric vector.}

\item{p}{Quantile probability between 0.5 and 1. Default = 0.75.}
}
\value{
Numeric right tail weight.
}
\description{
Computes a quantile-based measure of right tail weight.
}
