% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantiles.R
\name{hd.skew}
\alias{hd.skew}
\title{Quantile skewness}
\usage{
hd.skew(x, p = 0.25)
}
\arguments{
\item{x}{Numeric vector.}

\item{p}{Lower quantile probability, between 0 and 0.5. Default = 0.25.}
}
\value{
Numeric quantile skewness.
}
\description{
Computes a generalised quantile skewness measure, with default to Bowley's measure.
}
