% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantiles.R
\name{hd.uts}
\alias{hd.uts}
\title{Upper tail spread}
\usage{
hd.uts(x, q.C = 0.5, q.U = 0.9)
}
\arguments{
\item{x}{Numeric vector.}

\item{q.C}{Central quantile. Default = 0.5.}

\item{q.U}{Upper quantile. Default = 0.9.}
}
\value{
Numeric upper tail spread.
}
\description{
Computes the difference between an upper and central quantile.
}
