% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantiles.R
\name{hd_ineq}
\alias{hd_ineq}
\title{Upper-to-lower quantile ratio as a simple measure of inequality}
\usage{
hd_ineq(x, p = 0.1)
}
\arguments{
\item{x}{Numeric vector with positive lower quantile.}

\item{p}{Lower quantile probability, between 0 and 0.5. Default = 0.1.}
}
\value{
Numeric upper-to-lower quantile ratio.
}
\description{
Computes Q(1 - p) / Q(p), such as the 90/10 ratio.
}
\references{
Prendergast, L. A., Dedduwakumara, S., & Staudte, R. G. (2024).
rquest: An R package for hypothesis tests and confidence intervals for quantiles and summary measures based on quantiles
(arXiv:2410.11093; Version 1). arXiv. https://doi.org/10.48550/arXiv.2410.11093
}
