% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantiles.R
\name{hd_kurt}
\alias{hd_kurt}
\title{Moors quantile kurtosis}
\usage{
hd_kurt(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
Numeric Moors quantile kurtosis.
}
\description{
Computes the quantile-based kurtosis measure of Moors (1988).
}
\references{
Prendergast, L. A., Dedduwakumara, S., & Staudte, R. G. (2024).
rquest: An R package for hypothesis tests and confidence intervals for quantiles and summary measures based on quantiles
(arXiv:2410.11093; Version 1). arXiv. https://doi.org/10.48550/arXiv.2410.11093
}
