% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantiles.R
\name{hd_lts}
\alias{hd_lts}
\title{Lower tail spread}
\usage{
hd_lts(x, q.C = 0.5, q.L = 0.1)
}
\arguments{
\item{x}{Numeric vector.}

\item{q.C}{Central quantile. Default = 0.5.}

\item{q.L}{Lower quantile. Default = 0.1.}
}
\value{
Numeric lower tail spread.
}
\description{
Computes the difference between a central and lower quantile.
}
\references{
Prendergast, L. A., Dedduwakumara, S., & Staudte, R. G. (2024).
rquest: An R package for hypothesis tests and confidence intervals for quantiles and summary measures based on quantiles
(arXiv:2410.11093; Version 1). arXiv. https://doi.org/10.48550/arXiv.2410.11093
}
