% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantiles.R
\name{hd_niqr}
\alias{hd_niqr}
\title{Normalised interquantile range}
\usage{
hd_niqr(x, p = 0.25, q.L = 0.1, q.U = 0.9)
}
\arguments{
\item{x}{Numeric vector.}

\item{p}{Lower quantile probability, between 0 and 0.5. Default = 0.25.}

\item{q.L}{Lower reference quantile. Default = 0.1.}

\item{q.U}{Upper reference quantile. Default = 0.9.}
}
\value{
Numeric normalised interquantile range.
}
\description{
Computes (Q(1 - p) - Q(p)) / (Q(q.U) - Q(q.L)).
}
\references{
Prendergast, L. A., Dedduwakumara, S., & Staudte, R. G. (2024).
rquest: An R package for hypothesis tests and confidence intervals for quantiles and summary measures based on quantiles
(arXiv:2410.11093; Version 1). arXiv. https://doi.org/10.48550/arXiv.2410.11093
}
