% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantiles.R
\name{hd_qcd}
\alias{hd_qcd}
\title{Quantile coefficient of dispersion}
\usage{
hd_qcd(x, p = 0.25)
}
\arguments{
\item{x}{Numeric vector with positive quantiles.}

\item{p}{Lower quantile probability, between 0 and 0.5. Default = 0.25.}
}
\value{
Numeric quantile coefficient of dispersion.
}
\description{
Computes (Q(1 - p) - Q(p)) / (Q(1 - p) + Q(p)).
}
\references{
Prendergast, L. A., Dedduwakumara, S., & Staudte, R. G. (2024).
rquest: An R package for hypothesis tests and confidence intervals for quantiles and summary measures based on quantiles
(arXiv:2410.11093; Version 1). arXiv. https://doi.org/10.48550/arXiv.2410.11093
}
