% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantiles.R
\name{hd_rcv}
\alias{hd_rcv}
\title{Robust coefficient of variation}
\usage{
hd_rcv(x)
}
\arguments{
\item{x}{Numeric vector with a nonzero median.}
}
\value{
Numeric robust coefficient of variation.
}
\description{
Computes 0.75 * (Q(0.75) - Q(0.25)) / Q(0.5).
}
\references{
Prendergast, L. A., Dedduwakumara, S., & Staudte, R. G. (2024).
rquest: An R package for hypothesis tests and confidence intervals for quantiles and summary measures based on quantiles
(arXiv:2410.11093; Version 1). arXiv. https://doi.org/10.48550/arXiv.2410.11093
}
