% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantiles.R
\name{hd_rtw}
\alias{hd_rtw}
\title{Right tail weight}
\usage{
hd_rtw(x, p = 0.75)
}
\arguments{
\item{x}{Numeric vector.}

\item{p}{Quantile probability between 0.5 and 1. Default = 0.75.}
}
\value{
Numeric right tail weight.
}
\description{
Computes a quantile-based measure of right tail weight.
}
\references{
Prendergast, L. A., Dedduwakumara, S., & Staudte, R. G. (2024).
rquest: An R package for hypothesis tests and confidence intervals for quantiles and summary measures based on quantiles
(arXiv:2410.11093; Version 1). arXiv. https://doi.org/10.48550/arXiv.2410.11093
}
