% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantiles.R
\name{hd_uts}
\alias{hd_uts}
\title{Upper tail spread}
\usage{
hd_uts(x, q.C = 0.5, q.U = 0.9)
}
\arguments{
\item{x}{Numeric vector.}

\item{q.C}{Central quantile. Default = 0.5.}

\item{q.U}{Upper quantile. Default = 0.9.}
}
\value{
Numeric upper tail spread.
}
\description{
Computes the difference between an upper and central quantile.
}
\references{
Prendergast, L. A., Dedduwakumara, S., & Staudte, R. G. (2024).
rquest: An R package for hypothesis tests and confidence intervals for quantiles and summary measures based on quantiles
(arXiv:2410.11093; Version 1). arXiv. https://doi.org/10.48550/arXiv.2410.11093
}
