% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcoxmod.R
\name{hpsi}
\alias{hpsi}
\title{Evaluate Huber's Psi function for each value in the vector x}
\usage{
hpsi(x, bend = 1.28)
}
\arguments{
\item{x}{Numeric values.}

\item{bend}{Positive bending constant, with default to 1.28.}
}
\value{
The Huber Psi transformation of \code{x}.
}
\description{
Evaluate Huber's Psi function for each value in the vector x
}
\references{
Wilcox, Rand R. 2022. Introduction to Robust Estimation and Hypothesis Testing. 5th edn.
Statistical Modeling and Decision Science. San Diego, CA: Academic Press.
}
