% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcoxmod.R
\name{idealf}
\alias{idealf}
\title{Compute the ideal fourths for data in x}
\usage{
idealf(x)
}
\arguments{
\item{x}{Numeric data vector. Missing values must be removed first.}
}
\value{
A list containing the lower (\code{ql}) and upper (\code{qu}) ideal fourths.
}
\description{
Compute the ideal fourths for data in x
}
\references{
Carling, K. (2000). Resistant outlier rules and the non-Gaussian case.
Computational Statistics & Data Analysis, 33(3), 249-258. https://doi.org/10.1016/S0167-9473(99)00057-2
}
