% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{keeporder}
\alias{keeporder}
\title{Preserve the order of the names in a variable when converting to a factor}
\usage{
keeporder(x)
}
\arguments{
\item{x}{A vector to convert to a factor.}
}
\value{
A factor whose levels follow the order of first appearance in \code{x}.
}
\description{
Preserve the order of the names in a variable when converting to a factor
}
\examples{
# example code
df <- tibble::tibble(cond = c("A", "C", "B"))
df$cond <- keeporder(df$cond)
}
