% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcoxmod.R
\name{ks_ecdf}
\alias{ks_ecdf}
\title{Evaluate an empirical cumulative distribution function}
\usage{
ks_ecdf(x, val)
}
\arguments{
\item{x}{Numeric data vector.}

\item{val}{Value at which to evaluate the empirical distribution function.}
}
\value{
The proportion of values in \code{x} less than or equal to \code{val}.
}
\description{
Computes the proportion of values in \code{x} less than or equal to \code{val}.
This internal helper is used by \code{ks()}.
}
\keyword{internal}
