% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcoxmod.R
\name{mom}
\alias{mom}
\title{Modified one-step MOM location estimator}
\usage{
mom(x, bend = 2.24)
}
\arguments{
\item{x}{Numeric data vector.}

\item{bend}{Bending constant.}
}
\value{
A MOM estimate of location.
}
\description{
Modified one-step MOM location estimator
}
\references{
Wilcox, Rand R. 2022. Introduction to Robust Estimation and Hypothesis Testing. 5th edn.
Statistical Modeling and Decision Science. San Diego, CA: Academic Press.
}
