% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcoxmod.R
\name{mxmy}
\alias{mxmy}
\title{Estimate the median of the distribution of all pairwise differences.}
\usage{
mxmy(x, y, est = median, ...)
}
\arguments{
\item{x}{A numeric vector}

\item{y}{A numeric vector}

\item{est}{An estimator applied to the distribution X-Y, with default to the median.}

\item{...}{Additional arguments passed to \code{est}.}
}
\value{
A length-one object.
}
\description{
Estimate the median of the distribution of all pairwise differences.
}
\details{
The two vectors x and y are independent. The median is returned by default, but another estimator can be used.
}
\references{
Wilcox, Rand R. 2022. Introduction to Robust Estimation and Hypothesis Testing. 5th edn.
Statistical Modeling and Decision Science. San Diego, CA: Academic Press.

Hodges, J. L., & Lehmann, E. L. (1963). Estimates of Location Based on Rank Tests.
The Annals of Mathematical Statistics, 34(2), 598–611.
https://doi.org/10.1214/aoms/1177704172
}
\seealso{
\code{\link[=pxly]{pxly()}}
}
