% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcoxmod.R
\name{onestep}
\alias{onestep}
\title{One-step M-estimator of location}
\usage{
onestep(x, bend = 1.28, MED = TRUE)
}
\arguments{
\item{x}{Numeric data vector.}

\item{bend}{Huber's Psi bending constant, with default to 1.28.}

\item{MED}{Logical, with default to TRUE to use the median as the initial location estimate, otherwise use modified one-step M-estimator.}
}
\value{
A one-step M-estimate of location.
}
\description{
One-step M-estimator of location
}
\references{
Wilcox, Rand R. 2022. Introduction to Robust Estimation and Hypothesis Testing. 5th edn.
Statistical Modeling and Decision Science. San Diego, CA: Academic Press.
}
