% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_ecdf}
\alias{plot_ecdf}
\title{Plot one or two empirical cumulative distribution functions}
\usage{
plot_ecdf(
  x,
  y = NULL,
  xlab = "Measurements",
  ylab = "F(x) = proportion <= x",
  theme = NULL,
  x.name = "Group 1",
  y.name = "Group 2",
  x.col = "black",
  y.col = "grey60",
  x.lt = "longdash",
  y.lt = "solid",
  legend.width = 4,
  ...
)
}
\arguments{
\item{x}{Numeric vector for the first function.}

\item{y}{Optional numeric vector for the second function.}

\item{xlab, ylab}{Axis labels.}

\item{theme}{A ggplot2 theme, with default to \code{garstats::theme_gar()}.}

\item{x.name, y.name}{Labels for the plotted functions.}

\item{x.col, y.col}{Colours for the plotted functions.}

\item{x.lt, y.lt}{Line types for the plotted functions.}

\item{legend.width}{Width of the legend keys.}

\item{...}{Additional arguments passed to \code{ggplot2::stat_ecdf()}.}
}
\value{
A ggplot2 object.
}
\description{
Plot one or two empirical cumulative distribution functions
}
