% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcoxmod.R
\name{pxly}
\alias{pxly}
\title{Calculate P(X<Y), the probability that a random observation from vector x is less than a random observation from vector y.}
\usage{
pxly(x, y)
}
\arguments{
\item{x}{A numeric vector}

\item{y}{A numeric vector}
}
\value{
A probability \verb{[0,1]}.
}
\description{
Calculate P(X<Y), the probability that a random observation from vector x is less than a random observation from vector y.
}
\details{
The two vectors x and y are independent.
}
\references{
Wilcox, Rand R. 2022. Introduction to Robust Estimation and Hypothesis Testing. 5th edn.
Statistical Modeling and Decision Science. San Diego, CA: Academic Press.

Birnbaum, Z. W. (1956). On a Use of the Mann-Whitney Statistic.
In Proceedings of the Third Berkeley Symposium on Mathematical Statistics and Probability, Volume 1:
Contributions to the Theory of Statistics: 3.1 (pp. 13–18).
University of California Press.
}
\seealso{
\code{\link[=mxmy]{mxmy()}}
}
