% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcoxmod.R
\name{pxlyd}
\alias{pxlyd}
\title{Calculate P(X<Y) for paired observations}
\usage{
pxlyd(x, y, alpha = 0.05)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A numeric vector with the same length as \code{x}.}

\item{alpha}{Significance level for the (1-alpha)confidence interval, with 0.05
default.}
}
\value{
A list with the estimated probability (\code{phat}) and its confidence
interval (\code{ci}).
}
\description{
Estimates the probability that an observation in \code{x} is less than its paired
observation in \code{y}, with a confidence interval based on Pratt's method.
}
\details{
The vectors \code{x} and \code{y} contain dependent, paired observations.
}
\examples{
set.seed(666)
x <- rnorm(20)
y <- x + rnorm(20)
pxlyd(x, y)

}
\references{
Wilcox, Rand R. 2022. Introduction to Robust Estimation and Hypothesis Testing. 5th edn.
Statistical Modeling and Decision Science. San Diego, CA: Academic Press.
}
\seealso{
\code{\link[=pxly]{pxly()}}
}
