% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sim_counter}
\alias{sim_counter}
\title{Print simulation progress}
\usage{
sim_counter(S, nsim, inc, name = "Simulation")
}
\arguments{
\item{S}{Current simulation iteration.}

\item{nsim}{Total number of simulation iterations.}

\item{inc}{Frequency at which to print progress updates.}

\item{name}{Label printed with the initial progress message.}
}
\value{
Write progress to the console. No object is returned.
}
\description{
Print simulation progress
}
\examples{
n_iter <- 2000 # simulation iterations
inc.step <- 100 # increments
for (iter in 1:n_iter) {
  sim_counter(iter, n_iter, inc = inc.step, name = "Normal")
}
}
