% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcoxmod.R
\name{sint}
\alias{sint}
\title{Parametric confidence interval for a median}
\usage{
sint(x, y = NULL, alpha = 0.03, hyp = 0)
}
\arguments{
\item{x}{Numeric data vector for the first measurement. Missing values must
be removed first.}

\item{y}{Optional numeric data vector for the second measurement. When
supplied, it must have the same length as \code{x} and inference is performed
on \code{x - y}.}

\item{alpha}{Significance level, with default to 0.03 (97\% confidence).}

\item{hyp}{Null hypothesis value for the median, with 0 default.}
}
\value{
A list containing the median, sample size, confidence interval, and
p-value.
}
\description{
Uses the Hettmansperger-Sheather interpolation method.
}
\details{
If there are duplicate values in \code{x}, a gain in power might be obtained by calling \code{onesampb(x, est=hd)} instead.
This will provide a percentile bootstrap confidence interval for the Harrell-Davis estimator of the median.
}
\references{
Wilcox, Rand R. 2022. Introduction to Robust Estimation and Hypothesis Testing. 5th edn.
Statistical Modeling and Decision Science. San Diego, CA: Academic Press.
}
