% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcoxmod.R
\name{sint_ci}
\alias{sint_ci}
\title{Confidence interval for a median}
\usage{
sint_ci(x, alpha)
}
\arguments{
\item{x}{Numeric data vector.}

\item{alpha}{Significance level.}
}
\value{
A vector containing the lower and upper bounds of the confidence interval.
}
\description{
Confidence interval for a median
}
\keyword{internal}
