% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{theme_gar}
\alias{theme_gar}
\title{Package-specific ggplot2 theme}
\usage{
theme_gar(
  axis.title.size = 16,
  axis.text.size = 14,
  plot.title.size = 20,
  legend.text.size = 16,
  legend.title.size = 18,
  strip.text.size = 20,
  ...
)
}
\arguments{
\item{axis.title.size}{Font size for axis titles.}

\item{axis.text.size}{Font size for axis text.}

\item{plot.title.size}{Font size for plot titles.}

\item{legend.text.size}{Font size for legend text.}

\item{legend.title.size}{Font size for legend titles.}

\item{strip.text.size}{Font size for facet strip text.}

\item{...}{Additional arguments passed to \code{\link[ggplot2:theme]{ggplot2::theme()}}.}
}
\value{
A complete ggplot2 theme.
}
\description{
Package-specific ggplot2 theme
}
