% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcoxmod.R
\name{trimci}
\alias{trimci}
\title{Parametric inference for a one-sample trimmed mean}
\usage{
trimci(
  x,
  tr = 0.2,
  alpha = 0.03,
  hyp = 0,
  alternative = c("two.sided", "greater", "less")
)
}
\arguments{
\item{x}{Numeric data vector. Missing values must be removed first.}

\item{tr}{Trimming proportion in [0, 0.5), with 0.2 default.}

\item{alpha}{Significance level, with 0.03 default (97\% confidence).}

\item{hyp}{(Null) hypothesis value, with 0 default.}

\item{alternative}{Alternative hypothesis: "two.sided", "greater", or
"less", matching the alternative argument in t.test().}
}
\value{
A list containing the confidence interval, estimate, test statistic,
standard error, degrees of freedom, and p-value.
}
\description{
Parametric inference for a one-sample trimmed mean
}
\details{
For "greater" the confidence interval is \verb{[lower bound, Inf]}
For "less" the confidence interval is \verb{[-Inf, upper bound]}.
}
\examples{
n <- 20
set.seed(666)
# No effect: population trimmed mean = 0
x <- rnorm(20, mean = 0)
trimci(x, alternative = "two.sided")
trimci(x, alternative = "greater")
trimci(x, alternative = "less")
# Negative effect: population trimmed mean = -1
x <- rnorm(20, mean = -1)
trimci(x, alternative = "less")
# Positive effect: population trimmed mean = 1
x <- rnorm(20, mean = 1)
trimci(x, alternative = "greater")
# Inferences on means and 20\% trimmed means return very different results:
x <- rlnorm(50)
trimci(x, tr = 0, alternative = "less", hyp = 1.5)
trimci(x, tr = 0.2, alternative = "less", hyp = 1.5)
}
\references{
Wilcox, Rand R. 2022. Introduction to Robust Estimation and Hypothesis Testing. 5th edn.
Statistical Modeling and Decision Science. San Diego, CA: Academic Press.
}
