% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcoxmod.R
\name{trimse}
\alias{trimse}
\title{Parametric standard error of a trimmed mean}
\usage{
trimse(x, tr = 0.2)
}
\arguments{
\item{x}{Numeric data vector. Missing values must be removed first.}

\item{tr}{Trimming proportion in [0, 0.5), with default to 0.2.}
}
\value{
The estimated standard error of the trimmed mean.
}
\description{
Parametric standard error of a trimmed mean
}
\references{
Wilcox, Rand R. 2022. Introduction to Robust Estimation and Hypothesis Testing. 5th edn.
Statistical Modeling and Decision Science. San Diego, CA: Academic Press.
}
