% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcoxmod.R
\name{tsub}
\alias{tsub}
\title{Compute a bootstrap test statistic for paired trimmed means}
\usage{
tsub(isub, x, y, tr)
}
\arguments{
\item{isub}{Bootstrap sample indices.}

\item{x}{Numeric data vector for the first measurement.}

\item{y}{Numeric data vector for the second measurement.}

\item{tr}{Trimming proportion.}
}
\value{
A bootstrap test statistic.
}
\description{
Compute a bootstrap test statistic for paired trimmed means
}
\keyword{internal}
