% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcoxmod.R
\name{wincor.sub}
\alias{wincor.sub}
\title{Compute a Winsorized correlation for two numeric vectors}
\usage{
wincor.sub(x, y, tr = 0.2)
}
\arguments{
\item{x}{Numeric vector.}

\item{y}{Numeric vector of the same length as \code{x}.}

\item{tr}{Winsorization proportion in [0, 0.5), with 0.2 default.}
}
\value{
A list containing the Winsorized correlation, covariance, and p-value.
}
\description{
Compute a Winsorized correlation for two numeric vectors
}
\keyword{internal}
