% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcoxmod.R
\name{winvar}
\alias{winvar}
\title{Winsorized variance}
\usage{
winvar(x, tr = 0.2)
}
\arguments{
\item{x}{Numeric data vector. Missing values must be removed first.}

\item{tr}{Winsorization proportion in [0, 0.5).}
}
\value{
The Winsorized sample variance.
}
\description{
Winsorized variance
}
\references{
Wilcox, Rand R. 2022. Introduction to Robust Estimation and Hypothesis Testing. 5th edn.
Statistical Modeling and Decision Science. San Diego, CA: Academic Press.
}
