% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcoxmod.R
\name{yuend}
\alias{yuend}
\title{Parametric inference for two dependent trimmed means}
\usage{
yuend(x, y, tr = 0.2, alpha = 0.05)
}
\arguments{
\item{x}{Numeric data vector for the first measurement.}

\item{y}{Numeric data vector for the second measurement. Must have the same
length as \code{x}.}

\item{tr}{Trimming proportion in [0, 0.5), with 0.2 default.}

\item{alpha}{Significance level, with 0.05 default.}
}
\value{
A list containing the confidence interval, p-value, trimmed mean
estimates, difference in trimmed means, standard error, test statistic,
sample size, and degrees of freedom.
}
\description{
Compares the trimmed means of two paired samples using Yuen's method.
}
\details{
Missing values must be removed before calling this function. For inferences based on difference scores, use trimci() or trimcibt().
}
\examples{
set.seed(666)
x <- rnorm(20, mean = 1)
y <- x + rnorm(20)
yuend(x, y)

}
\references{
Wilcox, Rand R. 2022. Introduction to Robust Estimation and Hypothesis Testing. 5th edn.
Statistical Modeling and Decision Science. San Diego, CA: Academic Press.
}
\seealso{
\code{\link[=trimci]{trimci()}}, \code{\link[=trimcibt]{trimcibt()}}, \code{\link[=yuen]{yuen()}}
}
