% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcoxmod.R
\name{hd}
\alias{hd}
\title{Harrell-Davis quantile estimator}
\usage{
hd(x, q = 0.5)
}
\arguments{
\item{x}{Numeric data vector.}

\item{q}{Desired quantile, between 0 and 1.}
}
\value{
A numeric Harrell-Davis quantile estimate.
}
\description{
Harrell-Davis quantile estimator
}
\references{
Wilcox, Rand R. 2022. Introduction to Robust Estimation and Hypothesis Testing. 5th edn.
Statistical Modeling and Decision Science. San Diego, CA: Academic Press.

Harrell, F. E., & Davis, C. E. (1982). A new distribution-free quantile estimator.
Biometrika, 69(3), 635–640. https://doi.org/10.1093/biomet/69.3.635
}
