% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcoxmod.R
\name{sint}
\alias{sint}
\title{Parametric confidence interval for a median}
\usage{
sint(x, alpha = 0.03)
}
\arguments{
\item{x}{Numeric data vector. Missing values must be removed first.}

\item{alpha}{Significance level, with default to 0.03 (97\% confidence).}
}
\value{
A length-two numeric confidence interval.
}
\description{
Uses the Hettmansperger-Sheather interpolation method.
}
\details{
If there are duplicate values in \code{x}, a gain in power might be obtained by calling \code{onesampb(x, est=hd)} instead.
This will provide a percentile bootstrap confidence interval for the Harrell-Davis estimator of the median.
}
\references{
Wilcox, Rand R. 2022. Introduction to Robust Estimation and Hypothesis Testing. 5th edn.
Statistical Modeling and Decision Science. San Diego, CA: Academic Press.
}
